--- src/components/mentorship/MentorshipCalendar.css
@@ -1,69 +1,0 @@
-.mentorship-calendar {
-  padding: 24px;
-  background: white;
-  border-radius: 8px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-}
-
-.mentorship-calendar .events {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.mentorship-calendar .events li {
-  padding: 4px 8px;
-  margin-bottom: 4px;
-  background: #f5f5f5;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.mentorship-calendar .events li:hover {
-  background: #e6f7ff;
-}
-
-.mentorship-calendar .ant-badge-status {
-  width: 100%;
-  overflow: hidden;
-  font-size: 12px;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.mentorship-calendar .ant-badge-status-dot {
-  position: relative;
-  top: -1px;
-}
-
-.mentorship-calendar .ant-picker-calendar-full .ant-picker-panel {
-  background: #fff;
-  border: 1px solid #f0f0f0;
-}
-
-/* Modal styles */
-.ant-modal-content {
-  border-radius: 8px;
-}
-
-.ant-form-item-label > label {
-  font-weight: 500;
-}
-
-.ant-btn-primary {
-  height: 40px;
-  font-weight: 500;
-}
-
-/* Session details styles */
-.ant-descriptions-bordered .ant-descriptions-item-label {
-  background-color: #fafafa;
-  font-weight: 500;
-}
-
-.ant-tag {
-  border-radius: 4px;
-  padding: 4px 8px;
-  font-weight: 500;
-}
