+++ store.ts
@@ -7,7 +7,7 @@ import { notificationApi } from './src/services/notifications.ts'
 import { conversationApi } from './src/services/conversations.ts'
 import { inspirationApi } from './src/services/inspirations.ts'
 import { productApi } from './src/services/products.ts'
-import { mentorshipSessionsApi } from './src/services/mentorshipSessions.ts'
+import { mentorshipApi } from './src/services/mentorship.ts'
 
 export const store = configureStore({
     reducer: {
@@ -18,7 +18,7 @@ export const store = configureStore({
         [conversationApi.reducerPath]: conversationApi.reducer,
         [inspirationApi.reducerPath]: inspirationApi.reducer,
         [productApi.reducerPath]: productApi.reducer,
-        [mentorshipSessionsApi.reducerPath]: mentorshipSessionsApi.reducer
+        [mentorshipApi.reducerPath]: mentorshipApi.reducer
     },
     middleware: (getDefaultMiddleware) =>
         getDefaultMiddleware().concat(
@@ -29,7 +29,7 @@ export const store = configureStore({
             conversationApi.middleware,
             inspirationApi.middleware,
             productApi.middleware,
-            mentorshipSessionsApi.middleware
+            mentorshipApi.middleware
         ),
 })
